refactor _gen_row_access_intrinsic
